+++ resources/js/components/index.ts
@@ -8,3 +8,4 @@ export { default as Button } from './Button.vue';
 export { default as VoteButtons } from './VoteButtons.vue';
 export { default as HelpWantedSign } from './HelpWantedSign.vue';
 export { default as ToolCard } from './ToolCard.vue';
+export { default as SiteGuideModal } from './SiteGuideModal.vue';